// Frontend/src/app/models/maintenance-parameters.model.ts
export class MaintenanceParameters {
    device: string = "";
    description: string = "";
    outcome: string = "";
    status: string = "";
    scheduledDateStart: string = "";
    scheduledDateEnd: string = "";
    actualDateStart: string = "";
    actualDateEnd: string = "";
    createdAtStart: string = "";
    createdAtEnd: string = "";
    createdBy: string = "";
    pageNumber: number;
    pageSize: number;
    orderBy: string = "";
    orderDescending: boolean = false;
}